Move anvil tests into src and skip for now
